feat: create Access HistoryCard component
